add new module base_external_import to replace import_odbc
