+++ base_external_import/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_base_external_import_task_group_system,base_external_import_task_group_system,model_base_external_import_task,base.group_system,1,1,1,1
+access_base_external_import_log_group_system,base_external_import_log_group_system,model_base_external_import_log,,1,0,0,0
